Reorganize code
Move components that modify entities based on input to controls:js.
Move components that generate input events to inputs:js.
Move tests accordingly.
