+++ src/crafty.js
@@ -37,6 +37,7 @@ require('./graphics/webgl');
 require('./isometric/diamond-iso');
 require('./isometric/isometric');
 
+require('./controls/inputs');
 require('./controls/controls');
 require('./controls/device');
 require('./controls/keycodes');
