+++ tests/2d.js
@@ -118,8 +118,6 @@
 
   });
 
-
-
   test("circle", function() {
     var player = Crafty.e("2D").attr({
       w: 50,
@@ -251,6 +249,7 @@
   });
 
 
+  module("Collision");
 
   // This test assumes that the "circles" are really octagons, as per Crafty.circle.
   test("SAT overlap with circles", function() {
@@ -269,7 +268,6 @@
     strictEqual(e._SAT(c1, c2) !== false, true, "Polygons should test as overlapping");
 
   });
-
 
   test("adjustable boundary", function() {
     var e = Crafty.e("2D").attr({
@@ -300,197 +298,6 @@
     equal(e._by1, 5, "Y1 boundary set");
     equal(e._by2, 5, "Y2 boundary set");
 
-  });
-
-  test("Multiway", function() {
-    var e = Crafty.e("2D, Fourway")
-                  .attr({ x: 0, y: 0});
-
-    Crafty.trigger('KeyDown', {
-      key: Crafty.keys.W
-    });
-    Crafty.keydown[Crafty.keys.W] = true;
-    e.multiway(1, { W: -90 });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vy, -1);
-    equal(e._y, -1);
-
-    e.multiway(2, { W: 90 });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vy, 2);
-    equal(e._y, 1);
-
-    e.fourway(1);
-    Crafty.timer.simulateFrames(1);
-    equal(e._vy, -1);
-    equal(e._y, 0);
-
-    Crafty.timer.simulateFrames(1);
-    equal(e._vy, -1);
-    equal(e._y, -1);
-
-    
-    Crafty.trigger('KeyDown', {
-      key: Crafty.keys.UP_ARROW
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vy, -1);
-    equal(e._y, -2);
-
-    Crafty.trigger('KeyUp', {
-      key: Crafty.keys.W
-    });
-    delete Crafty.keydown[Crafty.keys.W];
-    Crafty.trigger('KeyUp', {
-      key: Crafty.keys.UP_ARROW
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vy, 0);
-    equal(e._y, -2);
-
-
-    Crafty.trigger('KeyDown', {
-      key: Crafty.keys.DOWN_ARROW
-    });
-    Crafty.trigger('KeyDown', {
-      key: Crafty.keys.LEFT_ARROW
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vy, 1);
-    equal(e._y, -1);
-    equal(e._vx, -1);
-    equal(e._x, -1);
-
-    Crafty.trigger('KeyUp', {
-      key: Crafty.keys.DOWN_ARROW
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vy, 0);
-    equal(e._y, -1);
-    equal(e._vx, -1);
-    equal(e._x, -2);
-
-    e.removeComponent("Multiway");
-    Crafty.timer.simulateFrames(1);
-    equal(e._vy, 0);
-    equal(e._y, -1);
-    equal(e._vx, 0);
-    equal(e._x, -2);
-
-    Crafty.trigger('KeyUp', {
-      key: Crafty.keys.LEFT_ARROW
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vy, 0);
-    equal(e._y, -1);
-    equal(e._vx, 0);
-    equal(e._x, -2);
-
-
-    e.destroy();
-  });
-
-  test("disableControl and enableControl and speed", function() {
-    var e = Crafty.e("2D, Twoway")
-      .attr({ x: 0 })
-      .twoway();
-
-    equal(e._vx, 0);
-    equal(e._x, 0);
-
-    e.enableControl();
-    e.speed({ x: 1, y: 1 });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 0);
-    equal(e._x, 0);
-
-    Crafty.trigger('KeyDown', {
-      key: Crafty.keys.D
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 1);
-    equal(e._x, 1);
-
-    e.disableControl();
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 0);
-    equal(e._x, 1);
-
-    Crafty.trigger('KeyUp', {
-      key: Crafty.keys.D
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 0);
-    equal(e._x, 1);
-
-    e.disableControl();
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 0);
-    equal(e._x, 1);
-
-
-    e.enableControl();
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 0);
-    equal(e._x, 1);
-
-    Crafty.trigger('KeyDown', {
-      key: Crafty.keys.D
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 1);
-    equal(e._x, 2);
-
-    Crafty.trigger('KeyUp', {
-      key: Crafty.keys.D
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 0);
-    equal(e._x, 2);
-
-
-    Crafty.trigger('KeyDown', {
-      key: Crafty.keys.D
-    });
-    Crafty.trigger('KeyDown', {
-      key: Crafty.keys.RIGHT_ARROW
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 1);
-    equal(e._x, 3);
-
-    e.disableControl();
-    e.speed({ x: 2, y: 2 });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 0);
-    equal(e._x, 3);
-
-    e.enableControl();
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 2);
-    equal(e._x, 5);
-
-    e.speed({ x: 3, y: 3 });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 3);
-    equal(e._x, 8);
-
-    Crafty.trigger('KeyUp', {
-      key: Crafty.keys.D
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 3);
-    equal(e._x, 11);
-
-    Crafty.trigger('KeyUp', {
-      key: Crafty.keys.RIGHT_ARROW
-    });
-    Crafty.timer.simulateFrames(1);
-    equal(e._vx, 0);
-    equal(e._x, 11);
-
-
-    e.destroy();
   });
 
 
@@ -631,6 +438,9 @@
     ok(e._cbr === null, "_cbr should now be removed along with Collision");
 
   });
+
+
+  module("Motion");
 
   test("Motion", function() {
     var Vector2D = Crafty.math.Vector2D;
