+++ tests/index.html
@@ -51,6 +51,7 @@
     <script type="text/javascript" src="./core.js"></script>
     <script type="text/javascript" src="./2d.js"></script>
     <script type="text/javascript" src="./audio.js"></script>
+    <script type="text/javascript" src="./inputs.js"></script>
     <script type="text/javascript" src="./controls.js"></script>
     <script type="text/javascript" src="./dom.js"></script>
     <script type="text/javascript" src="./events.js"></script>
